Update the doc, reduce the list of requirements
Update done in index.rst are to explain what and why the project is what it is.
Creation of install.rst is to help people setting up.
Reducing requirements in rtfd-requirements.txt is to avoid people installing useless package if they want to run the script.
It feel still bloated to me but at least it's a start.